rename IT tests to *Test.java
