+++ dspace/modules/dryad-widgets/dryad-widgets-webapp/src/test/java/it/BtnCiteLightboxTest.java
@@ -1,3 +1,4 @@
+
 package it;
 
 import java.util.List;
@@ -7,22 +8,22 @@ import static org.junit.Assert.*;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebElement;
 
-public class BtnZoomLightbox extends WidgetSeleniumTest {
+public class BtnCiteLightboxTest extends WidgetSeleniumTest {
 
   @Before
   public void setUp() throws Exception {
       super.setUp();
   }
+
   @After
   public void tearDown() throws Exception {
-      super.tearDown();
+    super.tearDown();
   }
 
   @Test
-  public void testBtnZoomLightbox() throws Exception {
-    String btn_selector = "a.dryad-ddw-zoom";
-    String zoom_header_selector = "#ddw-header";
-    String zoom_data_selector = "iframe.dryad-ddw-data";
+  public void testBtnCiteLightbox() throws Exception {
+    String btn_selector = "a.dryad-ddw-cite";
+    String citation_popup_selector = "#dryad-ddw-citation";
     driver.get(baseUrl + "/test.html");
     waitOnWidgetLoaded();
 
@@ -30,11 +31,9 @@ public class BtnZoomLightbox extends WidgetSeleniumTest {
     Boolean buttonWasClicked = clickFirstDisplayedInFrame(0, By.cssSelector(btn_selector));
     assertTrue(buttonWasClicked);
 
-    // confirm popup is open outside of iframe, and that the widget and
-    // content are present
+    // confirm quote content visible outer page
     waitUntilElementPresent(By.cssSelector(lightbox_container_selector), widgetPopupWaitSecondsTimeout);
-    assertTrue(isElementPresent(By.cssSelector(zoom_header_selector)));
-    assertTrue(isElementPresent(By.cssSelector(zoom_data_selector)));
+    assertTrue(isElementPresent(By.cssSelector(citation_popup_selector)));
 
     // out of frame
     assertTrue(isElementPresent(By.cssSelector(lightbox_close_selector)));
