+++ dspace/modules/dryad-widgets/dryad-widgets-webapp/src/test/java/it/BtnShareLightboxTest.java
@@ -1,41 +1,35 @@
-
 package it;
 
-import java.util.List;
 import junit.framework.TestCase;
 import org.junit.*;
 import static org.junit.Assert.*;
 import org.openqa.selenium.By;
-import org.openqa.selenium.WebElement;
 
-public class BtnCiteLightbox extends WidgetSeleniumTest {
+public class BtnShareLightboxTest extends WidgetSeleniumTest {
 
   @Before
   public void setUp() throws Exception {
       super.setUp();
   }
-
   @After
   public void tearDown() throws Exception {
     super.tearDown();
   }
 
+
   @Test
-  public void testBtnCiteLightbox() throws Exception {
-    String btn_selector = "a.dryad-ddw-cite";
-    String citation_popup_selector = "#dryad-ddw-citation";
+  public void testBtnShareLightbox() throws Exception {
+   
+    String btn_selector = "i.fa.fa-share-alt:nth-of-type(1)";
+    String close_selector = "button.mfp-close";
     driver.get(baseUrl + "/test.html");
     waitOnWidgetLoaded();
-
+    
     // click button in widget frame
     Boolean buttonWasClicked = clickFirstDisplayedInFrame(0, By.cssSelector(btn_selector));
     assertTrue(buttonWasClicked);
-
-    // confirm quote content visible outer page
-    waitUntilElementPresent(By.cssSelector(lightbox_container_selector), widgetPopupWaitSecondsTimeout);
-    assertTrue(isElementPresent(By.cssSelector(citation_popup_selector)));
-
-    // out of frame
+    
+    // close lightbox
     assertTrue(isElementPresent(By.cssSelector(lightbox_close_selector)));
     driver.findElement(By.cssSelector(lightbox_close_selector)).click();
     waitUntilElementAbsent(By.cssSelector(lightbox_close_selector),widgetLoadedSecondsTimeout);
