+++ dspace/modules/dryad-widgets/dryad-widgets-webapp/src/test/java/it/BtnZoomLightboxTest.java
@@ -1,11 +1,13 @@
 package it;
 
+import java.util.List;
 import junit.framework.TestCase;
 import org.junit.*;
 import static org.junit.Assert.*;
 import org.openqa.selenium.By;
+import org.openqa.selenium.WebElement;
 
-public class BtnShareLightbox extends WidgetSeleniumTest {
+public class BtnZoomLightboxTest extends WidgetSeleniumTest {
 
   @Before
   public void setUp() throws Exception {
@@ -13,23 +15,28 @@ public class BtnShareLightbox extends WidgetSeleniumTest {
   }
   @After
   public void tearDown() throws Exception {
-    super.tearDown();
+      super.tearDown();
   }
 
-
   @Test
-  public void testBtnShareLightbox() throws Exception {
-   
-    String btn_selector = "i.fa.fa-share-alt:nth-of-type(1)";
-    String close_selector = "button.mfp-close";
+  public void testBtnZoomLightbox() throws Exception {
+    String btn_selector = "a.dryad-ddw-zoom";
+    String zoom_header_selector = "#ddw-header";
+    String zoom_data_selector = "iframe.dryad-ddw-data";
     driver.get(baseUrl + "/test.html");
     waitOnWidgetLoaded();
-    
+
     // click button in widget frame
     Boolean buttonWasClicked = clickFirstDisplayedInFrame(0, By.cssSelector(btn_selector));
     assertTrue(buttonWasClicked);
-    
-    // close lightbox
+
+    // confirm popup is open outside of iframe, and that the widget and
+    // content are present
+    waitUntilElementPresent(By.cssSelector(lightbox_container_selector), widgetPopupWaitSecondsTimeout);
+    assertTrue(isElementPresent(By.cssSelector(zoom_header_selector)));
+    assertTrue(isElementPresent(By.cssSelector(zoom_data_selector)));
+
+    // out of frame
     assertTrue(isElementPresent(By.cssSelector(lightbox_close_selector)));
     driver.findElement(By.cssSelector(lightbox_close_selector)).click();
     waitUntilElementAbsent(By.cssSelector(lightbox_close_selector),widgetLoadedSecondsTimeout);
