+++ dspace/modules/dryad-widgets/dryad-widgets-webapp/src/test/java/it/WidgetLoadTest.java
@@ -5,7 +5,7 @@ import org.junit.*;
 import static org.junit.Assert.*;
 import org.openqa.selenium.By;
 
-public class ButtonsPresent extends WidgetSeleniumTest {
+public class WidgetLoadTest extends WidgetSeleniumTest {
 
   @Before
   public void setUp() throws Exception {
@@ -17,15 +17,11 @@ public class ButtonsPresent extends WidgetSeleniumTest {
   }
 
   @Test
-  public void testButtonsPresent() throws Exception {
+  public void testWidgetLoad() throws Exception {
     driver.get(baseUrl + "/test.html");
     waitOnWidgetLoaded();
-
+    
     driver.switchTo().frame(0);
-    assertTrue(isElementPresent(By.cssSelector("i.fa.fa-expand")));
-    assertTrue(isElementPresent(By.cssSelector("i.fa.fa-share-alt")));
-    assertTrue(isElementPresent(By.cssSelector("i.fa.fa-download")));
-    assertTrue(isElementPresent(By.cssSelector("i.fa.fa-quote-left")));
+    assertTrue(isElementPresent(By.cssSelector("img[alt=\"Data in Dryad\"]")));
   }
-
 }
